CEDS-2975 Refactoring
