+++ test/unit/uk/gov/hmrc/exports/models/declaration/notifications/NotificationSpec.scala
@@ -20,14 +20,16 @@ import java.time.format.DateTimeFormatter
 import java.time.{LocalDateTime, ZoneId, ZonedDateTime}
 
 import play.api.libs.json.Json
+import testdata.ExportsTestData.{actionId, mrn}
+import testdata.notifications.ExampleXmlAndNotificationDetailsPair.exampleReceivedNotification
 import uk.gov.hmrc.exports.base.UnitSpec
 import uk.gov.hmrc.exports.models.declaration.submissions.SubmissionStatus
 
-class NotificationsSpec extends UnitSpec {
+class NotificationSpec extends UnitSpec {
 
   val formatter = DateTimeFormatter.ISO_ZONED_DATE_TIME
 
-  "Notifications Spec" must {
+  "Notification" must {
     val actionId = "123"
     val dateTime = LocalDateTime.now()
     val mrn = "id1"
@@ -42,7 +44,7 @@ class NotificationsSpec extends UnitSpec {
     "have json writes that produce object which could be parsed by the front end service" in {
       val json = Json.toJson(notification)(Notification.FrontendFormat.writes)
 
-      json.toString() mustBe NotificationsSpec.serialisedWithNonOptionalDetailsFormat(
+      json.toString() mustBe NotificationSpec.serialisedWithNonOptionalDetailsFormat(
         actionId,
         mrn,
         dateTime.atZone(ZoneId.of("UCT")).format(formatter),
@@ -53,7 +55,7 @@ class NotificationsSpec extends UnitSpec {
     "have json writes that produce object which could be parsed by the database" in {
       val json = Json.toJson(notification)(Notification.DbFormat.writes)
 
-      json.toString() mustBe NotificationsSpec.serialisedWithOptionalDetailsFormat(
+      json.toString() mustBe NotificationSpec.serialisedWithOptionalDetailsFormat(
         actionId,
         mrn,
         dateTime.atZone(ZoneId.of("UCT")).format(formatter),
@@ -61,9 +63,35 @@ class NotificationsSpec extends UnitSpec {
       )
     }
   }
+
+  "Notification on unparsed" should {
+
+    val xml = exampleReceivedNotification(mrn).asXml
+
+    "return Notification with actionId" in {
+
+      val result = Notification.unparsed(actionId, xml)
+
+      result.actionId mustBe actionId
+    }
+
+    "return Notification with payload" in {
+
+      val result = Notification.unparsed(actionId, xml)
+
+      result.payload mustBe xml.toString
+    }
+
+    "return Notification with empty details" in {
+
+      val result = Notification.unparsed(actionId, xml)
+
+      result.details mustBe empty
+    }
+  }
 }
 
-object NotificationsSpec {
+object NotificationSpec {
   def serialisedWithNonOptionalDetailsFormat(actionId: String, mrn: String, dateTime: String, payload: String) =
     s"""{"actionId":"${actionId}","mrn":"${mrn}","dateTimeIssued":"${dateTime}","status":"ACCEPTED","errors":[],"payload":"${payload}"}"""
